Put event management content in events management page
